# Conflicts:
#	Fakturama-Parent/com.sebulli.fakturama.rcp/src/main/java/com/sebulli/fakturama/parts/ContactEditor.java
#	Fakturama-Parent/com.sebulli.fakturama.rcp/src/main/java/com/sebulli/fakturama/parts/Editor.java
#	Fakturama-Parent/com.sebulli.fakturama.rcp/src/main/java/com/sebulli/fakturama/parts/ProductEditor.java
#	Fakturama-Parent/com.sebulli.fakturama.rcp/src/main/java/com/sebulli/fakturama/webshopimport/WebShopDataImporter.java
#	Fakturama-Parent/com.sebulli.fakturama.site/com.sebulli.fakturama.product
#	Fakturama-Parent/org.fakturama.target/org.fakturama.target.target
#	Fakturama-Parent/pom.xml

+++ Fakturama-Parent/com.sebulli.fakturama.rcp/src/main/java/com/sebulli/fakturama/parts/voucheritems/VoucherItemListBuilder.java
@@ -101,7 +101,7 @@ public class VoucherItemListBuilder {
 
             // Add a new item with default properties
             public void mouseDown(MouseEvent e) {
-                int defaultVatId = preferences.getInt(Constants.DEFAULT_VAT, 1);
+                long defaultVatId = preferences.getLong(Constants.DEFAULT_VAT, 1);
                 // Use the standard VAT value
                 VAT defaultVat = vatsDao.findById(defaultVatId);
                 VoucherItem item = createNewVoucherItem(msg.commonFieldName, "", Double.valueOf(0.0), defaultVat, container.getVoucherType());
